modules/CANDataHandler/requirements.txt
azure-iot-device~=2.0.0
six~=1.15.0